+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@ Release Notes
 
 ## 1.6.21
 * CDN: Adds new SKU types for Azure Front Door Standard/Premium
+* Support for GPUs in Azure Container Instances
 
 ## 1.6.20
 * CDN rules: Only make CacheDuration required for Override and SetIfMissing and not BypassCache when creating cache_expiration action