+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@ Release Notes
 
 ## 1.6.21
 * CDN: Adds new SKU types for Azure Front Door Standard/Premium
+* Key Vaults: Fixed bug where adding vnetRules to KeyVault did not work.
 * Support for GPUs in Azure Container Instances
 
 ## 1.6.20