+++ RELEASE_NOTES.md
@@ -1,12 +1,11 @@
 Release Notes
 =============
-## vNext
-* Functions: fix for .Net isolated functions hosted on linux
 
 ## 1.6.21
-* Alerts: Extend a list of possible criteria's time aggregations and operators
+* Alerts: Extend a list of possible criteria for time aggregations and operators
 * Alerts: Support of custom metric alerts
 * CDN: Adds new SKU types for Azure Front Door Standard/Premium
+* Functions: Fix for .NET isolated functions hosted on Linux
 * Key Vaults: Fixed bug where adding vnetRules to KeyVault did not work.
 * Support for GPUs in Azure Container Instances